+++ src/lib/index.js
@@ -4,7 +4,7 @@
  * Created Date: 2018-03-26, 13:16:18
  * Author: Przemysław Drzewicki <przemyslaw.drzewicki@gmail.com>
  * =============================================================================
- * Last Modified: 2018-03-28, 11:02:11
+ * Last Modified: 2018-03-28, 11:41:42
  * Modified By: Przemysław Drzewicki
  * =============================================================================
  * Copyright (c) 2018 webonweb
@@ -41,10 +41,11 @@ export default class DragList extends React.Component {
    * @memberof DragList
    */
   state = {
-    current: null,
     target: null,
-    placeholder: true,
-    handler: true
+    current: null,
+    handler: true,
+    propagation: false,
+    placeholder: true
   }
 
   /**
@@ -99,7 +100,7 @@ export default class DragList extends React.Component {
     this.props.list.map((item, index) => (
       <div key={`drag-list-${index}`}
         className={this.globals.mainClassName}
-        onMouseDown={this.state.handler? this.handleMouseDown : () => {}}
+        onMouseDown={!this.state.handler? this.handleMouseDown : () => {}}
         onMouseUp={this.handleMouseUp}  
         onMouseMove={this.handleMouseMove}
         data-index={index}
@@ -123,8 +124,34 @@ export default class DragList extends React.Component {
    * @memberof DragList
    */
   handleMouseMove = (event = {}) => {
+    if(this.state.propagation) { 
+      const parent = event.target.parentElement;
+      if(parent.className !== this.globals.mainClassName) { return false; }
+      this.startHandleDrag(parent, event);
+    }
+
     if(!this.state.current) { return false; }
     this.setState({ target: event.currentTarget.dataset.index })
+  }
+
+  /**
+   * Start propagation element drag.
+   * 
+   * @param {any} [parent={}] 
+   * @param {any} [event={}] 
+   * @memberof DragList
+   */
+  startHandleDrag = (parent = {}, event = {}) => {
+    const node = this.state.handler? nodeToString(parent).replace('<div class="drag-list-handler"><span></span><span></span><span></span></div>', '')
+      : nodeToString(event.target);
+
+    this.setState({
+      current: {
+        index: this.state.handler? parent.dataset.index: event.currentTarget.dataset.index,
+        element: <div dangerouslySetInnerHTML={{ __html: node }} />
+      },
+      propagation: false
+    });
   }
 
   /**
@@ -133,18 +160,8 @@ export default class DragList extends React.Component {
    * @param {any} [event={}] 
    * @memberof DragList
    */
-  handleMouseDown = (event = {}) => {
-    const parent = event.target.parentElement;
-    if(parent.className !== this.globals.mainClassName) { return false; }
-    
-    const node = nodeToString(parent).replace('<div class="drag-list-handler"><span></span><span></span><span></span></div>', '');
-    this.setState({
-      current: {
-        index: parent.dataset.index,
-        element: <div dangerouslySetInnerHTML={{ __html: node }} />
-      } 
-    })
-  }
+  handleMouseDown = (event = {}) => 
+    this.setState({ propagation: true });
 
   /**
    * Handle mouse up event.
@@ -153,16 +170,27 @@ export default class DragList extends React.Component {
    * @memberof DragList
    */
   handleMouseUp = (event = {}) => {
-    if(!this.state.current) { return false; }
+    if(!this.state.current) { 
+      this.clear();
+      return false; 
+    }
     const index = event.currentTarget.dataset.index? event.currentTarget.dataset.index : this.state.target;
-    if(this.state.current.index === index){
-      this.setState({ current: null, target: null });
+    if(this.state.current.index === index || !index || !this.state.current.index){
+      this.clear();
       return false;
     }
     this.props.update(this.props.list[this.state.current.index], this.state.current.index, index);
-    this.setState({ current: null, target: null });
+    this.setState({ current: null, target: null, propagation: false });
   }
 
+  /**
+   * Clear all state.
+   * 
+   * @memberof DragList
+   */
+  clear = () => 
+    this.setState({ current: null, target: null, propagation: false });
+    
   /**
    * Main rendering function.
    *