+++ src/lib/template.js
@@ -4,7 +4,7 @@
  * Created Date: 2018-03-26, 13:21:06
  * Author: Przemysław Drzewicki <przemyslaw.drzewicki@gmail.com>
  * =============================================================================
- * Last Modified: 2018-03-28, 11:02:14
+ * Last Modified: 2018-03-28, 11:13:47
  * Modified By: Przemysław Drzewicki
  * =============================================================================
  * Copyright (c) 2018 webonweb